SolverEngine Refactor Phase 2: Introduce SolveStrategy.
- Add SolveStrategy class responsible for concrete solving strategy. Decoupliing strategy implementations from SolverEngine.
- Adapt PrintUtil, Dataflow type system to this change.
