+++ src/checkers/inference/solver/backend/strategy/SolveStrategy.java
@@ -1,0 +1,19 @@
+package checkers.inference.solver.backend.strategy;
+
+import java.util.Collection;
+import java.util.Map;
+
+import javax.annotation.processing.ProcessingEnvironment;
+
+import org.checkerframework.framework.type.QualifierHierarchy;
+
+import checkers.inference.InferenceSolution;
+import checkers.inference.model.Constraint;
+import checkers.inference.model.Slot;
+
+public interface SolveStrategy {
+
+    InferenceSolution solve(Map<String, String> configuration, Collection<Slot> slots,
+            Collection<Constraint> constraints, QualifierHierarchy qualHierarchy,
+            ProcessingEnvironment processingEnvironment);
+}
